Many changes, mostly to the way dsp_examples are run (now via portaudio's API callback and the new C-API for dclang)

+++ dclang.h
@@ -4,6 +4,6 @@ extern int dclang_import(char *infilestr);
 
 extern int dclang_findword(const char *word);
 
-extern void dclang_callword(double where);
+extern void dclang_callword(unsigned long int where);
 
 extern double dclang_pop();
